feat: :card_file_box: データクラスをdrift, freezed両対応に
