+++ lib/src/features/movies/repository/movie_repository.dart
@@ -2,10 +2,11 @@ import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 import 'package:tmdb_app/env.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
-import 'package:tmdb_app/src/features/movies/dao/database.dart';
+
 import 'package:tmdb_app/src/features/movies/data_model/movie_response/movie/movie.dart';
 import 'package:tmdb_app/src/features/movies/data_model/movie_response/movie_response.dart';
 import 'package:tmdb_app/src/features/movies/data_model/review_response/review_response.dart';
+import 'package:tmdb_app/src/features/movies/database/local_database.dart';
 
 import 'package:tmdb_app/src/utils/dio/dio_provider.dart';
 part 'movie_repository.g.dart';
